Reorganise some code
Mostly in preparation for the next but one commit.

+++ sigbak.h
@@ -179,7 +179,7 @@ void		 sbk_free_file(struct sbk_file *);
 struct sbk_attachment_list *sbk_get_all_attachments(struct sbk_ctx *);
 struct sbk_attachment_list *sbk_get_attachments_for_mms(struct sbk_ctx *, int);
 struct sbk_attachment_list *sbk_get_attachments_for_thread(struct sbk_ctx *,
-		    uint64_t);
+		    int);
 void		 sbk_free_attachment_list(struct sbk_attachment_list *);
 
 struct sbk_sms_list *sbk_get_smses(struct sbk_ctx *, int);
